requirements.txt
streamlit
openai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pandas
python-docx
reportlab
PyPDF2
opencv-python-headless
Pillow
odfpy
